Create separate module for subquery logic
